Move everything specific to the current api to legacy
