+++ waspc/data/packages/wasp-config/__tests__/legacy/testFixtures.ts
@@ -3,10 +3,10 @@
  * In our case the sample data represents TsAppSpec data.
  */
 
-import * as AppSpec from "../src/appSpec.js";
-import { Branded } from "../src/branded.js";
-import { App } from "../src/publicApi/legacy/App.js";
-import * as TsAppSpec from "../src/publicApi/legacy/tsAppSpec.js";
+import * as AppSpec from "../../src/appSpec.js";
+import { Branded } from "../../src/branded.js";
+import { App } from "../../src/legacy/publicApi/App.js";
+import * as TsAppSpec from "../../src/legacy/publicApi/tsAppSpec.js";
 
 export function createApp(scope: ConfigType): {
   appConfigName: string;
