+++ waspc/data/packages/wasp-config/src/legacy/publicApi/tsAppSpec.ts
